FetchDeviceOperation: account for pagination
