+++ Sources/XKit/DeveloperServices/DeveloperServicesProvisioningOperation.swift
@@ -39,7 +39,7 @@ public struct DeveloperServicesProvisioningOperation: DeveloperServicesOperation
 
     public func perform() async throws -> Response {
         progress(0/3)
-        _ = try await DeveloperServicesFetchDeviceOperation(context: context).perform()
+        try await DeveloperServicesAddDeviceOperation(context: context).perform()
 
         progress(1/3)
         let signingInfo = try await DeveloperServicesFetchCertificateOperation(
